app/requirements.txt
flask
flask_cors
pytest
mysql-connector



